Move some waiter code into dedicated files
