+++ pkg/operation/botanist/waiter.go
@@ -131,13 +131,6 @@ func (b *Botanist) WaitUntilKubeAPIServerReady(ctx context.Context) error {
 	return nil
 }
 
-// WaitForKubeControllerManagerToBeActive waits for the kube controller manager of a Shoot cluster has acquired leader election, thus is active.
-func (b *Botanist) WaitForKubeControllerManagerToBeActive(ctx context.Context) error {
-	b.Shoot.Components.ControlPlane.KubeControllerManager.SetShootClient(b.K8sShootClient.Client())
-
-	return b.Shoot.Components.ControlPlane.KubeControllerManager.WaitForControllerToBeActive(ctx)
-}
-
 // WaitUntilTunnelConnectionExists waits until a port forward connection to the tunnel pod (vpn-shoot or konnectivity-agent) in the kube-system
 // namespace of the Shoot cluster can be established.
 func (b *Botanist) WaitUntilTunnelConnectionExists(ctx context.Context) error {
@@ -148,25 +141,6 @@ func (b *Botanist) WaitUntilTunnelConnectionExists(ctx context.Context) error {
 		}
 
 		return b.CheckTunnelConnection(ctx, b.Logger, tunnelName)
-	})
-}
-
-// WaitUntilSeedNamespaceDeleted waits until the namespace of the Shoot cluster within the Seed cluster is deleted.
-func (b *Botanist) WaitUntilSeedNamespaceDeleted(ctx context.Context) error {
-	return b.waitUntilNamespaceDeleted(ctx, b.Shoot.SeedNamespace)
-}
-
-// WaitUntilNamespaceDeleted waits until the <namespace> within the Seed cluster is deleted.
-func (b *Botanist) waitUntilNamespaceDeleted(ctx context.Context, namespace string) error {
-	return retry.UntilTimeout(ctx, 5*time.Second, 900*time.Second, func(ctx context.Context) (done bool, err error) {
-		if err := b.K8sSeedClient.Client().Get(ctx, client.ObjectKey{Name: namespace}, &corev1.Namespace{}); err != nil {
-			if apierrors.IsNotFound(err) {
-				return retry.Ok()
-			}
-			return retry.SevereError(err)
-		}
-		b.Logger.Infof("Waiting until the namespace '%s' has been cleaned up and deleted in the Seed cluster...", namespace)
-		return retry.MinorError(fmt.Errorf("namespace %q is not yet cleaned up", namespace))
 	})
 }
